backed_django/requirements.txt
Django>=4.2,<5
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
psycopg2-binary==2.9.10
gunicorn==21.2.0
django-cors-headers==4.0.0
django-environ==0.12.0


